requirements.txt
# Core Framework
streamlit==1.48.1

# Machine Learning & Data Science
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Image Processing
Pillow==10.0.0

# Visualization
plotly==5.17.0

# Database
pymongo==4.5.0
sqlalchemy==2.0.21
python-dotenv==1.0.0

# Security
bcrypt==4.0.1

# PDF Generation
fpdf==1.7.2

# Optional: For GPU support (uncomment if needed)
# tensorflow-gpu==2.15.0

# Development Tools (optional)
# pytest==7.4.2
# black==23.9.1
# flake8==6.1.0
# mypy==1.5.1
